requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorthief==0.2.1
Django==4.1.2
idna==3.4
Pillow==9.2.0
pip==22.2.2
requests==2.28.1
setuptools==65.4.1
sqlparse==0.4.3
urllib3==1.26.12
wheel==0.37.1
